Add data.csv, submission.csv and random_search.ipynb notebbok
data.csv is the output of data_process.csv and its used to train the ml models.
submission.csv contains the rows that did not have a value for total_minutes.
random_search.ipynb preforms a randomized search to find good parameters for three regressors (SVR, RFR and XGBRegressor) according to the mae metric.